[TASK] refinement of price value inputs of filter
Because the prices must be passed as floats, but the user may input
arbitrary formatted price value, the Filter domain object now comes with
a new price value normalization method for this kind of data. As PHP 5.3
ships with the intl package, this can be used but we need 5.2
compatibility as TYPO3 4.5 supports this PHP version.

fixed #1